Frontend/requirements.txt
streamlit
requests
pillow
boto3
python-dotenv